requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
dill==0.3.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-resources==3.3.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tb-nightly==2.5.0a20201121
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==4.1.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020102301
tf-nightly==2.5.0.dev20201121
tfds-nightly==4.1.0.dev202012080107
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
